Done category status priority

+++ lib/db/CategoryHelper.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/cupertino.dart';
 import 'package:note_management_system/db/AppSQLHelper.dart';
 import 'package:note_management_system/model/Categories.dart';
 // ignore: depend_on_referenced_packages
@@ -16,6 +15,26 @@ class CategoryHelper {
     } else {
       return null;
     }
+  }
+
+  static Future<int?> updateItem(Categories categories) async {
+    final db = await AppSQLHelper.db();
+    if(await checkCategoryAvailableByNameAndUserID(categories.name!, categories.userId!)){
+      return  await db.update(Constant.KEY_TABLE_CATEGORY, categories.toMap(),
+          where: 'id = ?', whereArgs: [categories.id]);
+    } else {
+      return null;
+    }
+  }
+
+  static Future<bool> checkCategoryAvailableByNameAndUserID(String name, int userId) async {
+    final db = await AppSQLHelper.db();
+    final List<Map<String, dynamic>> maps = await db
+        .query(Constant.KEY_TABLE_CATEGORY,
+        where: '${Constant.KEY_CATEGORY_NAME} = ? AND ${Constant.KEY_CATEGORY_USER_ID} = ?',
+        whereArgs: [name, userId]);
+
+    return maps.isEmpty;
   }
 
   static Future<List<Map<String, dynamic>>> getAllItem(int userId) async {
@@ -38,33 +57,24 @@ class CategoryHelper {
     }
   }
 
-  static Future<int> updateItem(Categories categories) async {
+  static Future<bool> checkCategoryInUse(int id) async {
     final db = await AppSQLHelper.db();
-    final result = await db.update(Constant.KEY_TABLE_CATEGORY, categories.toMap(),
-        where: '${Constant.KEY_CATEGORY_ID} = ?', whereArgs: [categories.id]);
-    return result;
+    final note = await db.query(
+      Constant.KEY_TABLE_NOTE,
+      where: '${Constant.KEY_NOTE_CATEGORY_ID} = ?',
+      whereArgs: [id],
+    );
+    return note.isNotEmpty;
   }
 
   static Future<int?> deleteItem(int id) async {
     final db = await AppSQLHelper.db();
 
-    try {
+    if(await checkCategoryInUse(id)){
+      return null;
+    } else {
       return await db.delete(Constant.KEY_TABLE_CATEGORY,
           where: '${Constant.KEY_CATEGORY_ID} = ?', whereArgs: [id]);
-    } catch (err) {
-      debugPrint("Something went wrong when deleting an item: $err");
     }
-    return null;
   }
-
-  static Future<bool> checkCategoryAvailableByNameAndUserID(String name, int userId) async {
-    final db = await AppSQLHelper.db();
-    final List<Map<String, dynamic>> maps = await db
-        .query(Constant.KEY_TABLE_CATEGORY,
-        where: '${Constant.KEY_CATEGORY_NAME} = ? AND ${Constant.KEY_CATEGORY_USER_ID} = ?',
-        whereArgs: [name, userId]);
-
-    return maps.isEmpty;
-  }
-
 }
